+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - Unreleased
+
+### Added
+
+- `datacube` `mask_with` parameters supports now `cloud_mask` (cnn cloudmask from EDA).
+
+### Fixed 
+
+- items `Datetime` are automatically converted in nanoseconds.
+- AssetMapper done also in search for only requestins the assets
+
 ## [0.1.7] - 2024-06-14
 
 ### Fixed